requirement.txt
amqp==2.3.2
aniso8601==3.0.2
asn1crypto==0.24.0
billiard==3.5.0.4
celery==4.2.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.2.2
Flask==1.0.2
Flask-JWT-Extended==3.10.0
Flask-MySQL==1.4.0
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
kombu==4.2.1
MarkupSafe==1.0
mysqlclient==1.3.13
pycparser==2.18
PyJWT==1.6.4
PyMySQL==0.9.2
python-crontab==2.3.4
python-dateutil==2.7.3
pytz==2018.4
requests==2.19.1
six==1.11.0
SQLAlchemy==1.2.8
urllib3==1.23
vine==1.1.4
Werkzeug==0.14.1
